requirements.txt
# LangChain / LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-qdrant>=0.2.0
langgraph>=0.2.0

# Vector DB
qdrant-client>=1.9.0

# LLM 및 임베딩
openai>=1.30.0
sentence-transformers>=3.0.0
torch>=2.0.0

# PDF 처리
pypdf>=4.0.0
langchain-text-splitters>=0.3.0

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
